refactor: organize base models
