+++ geoips/pydantic_models/v1/workflows.py
@@ -25,12 +25,12 @@ from pydantic import ConfigDict, Field, field_validator, model_validator
 
 # GeoIPS imports
 from geoips import interfaces
-from geoips.pydantic_models.v1.bases import (
+from geoips.pydantic_models.bases import (
     PythonIdentifier,
-    PluginModel,
     FrozenModel,
     PermissiveFrozenModel,
 )
+from geoips.pydantic_models.v1.bases import PluginModel
 from geoips.utils.types.partial_lexeme import Lexeme
 
 LOG = logging.getLogger(__name__)
